+++ client/src/features/moment/const/tags.ts
@@ -1,0 +1,1 @@
+export const TAGS = ['일상/생각', '인간관계', '일/성장', '건강/운동', '취미/여가'];
