+++ client/src/shared/ui/tag/TagList.tsx
@@ -2,13 +2,12 @@ import styled from '@emotion/styled';
 import Tag from './Tag';
 
 interface TagListProps {
+  tags: string[];
   onTagClick: (tagName: string) => void;
   selectedTag?: string[];
 }
 
-export const TagList = ({ onTagClick, selectedTag }: TagListProps) => {
-  const tags = ['일상/생각', '인간관계', '일/성장', '건강/운동', '취미/여가'];
-
+export const TagList = ({ tags, onTagClick, selectedTag }: TagListProps) => {
   return (
     <TagListStyles>
       {tags.map(tag => (
